modules/sagemaker/sagemaker-groundtruth/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
attrs==23.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.202
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.3
    # via aws-cdk-lib
aws-cdk-aws-lambda-python-alpha==2.147.3a0
    # via -r requirements.in
aws-cdk-lib==2.147.3
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   cdk-nag
boto3==1.21.46
    # via -r requirements.in
botocore==1.24.46
    # via
    #   boto3
    #   s3transfer
cattrs==23.2.3
    # via jsii
cdk-nag==2.12.29
    # via -r requirements.in
constructs==10.0.91
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
exceptiongroup==1.2.1
    # via cattrs
importlib-resources==6.4.0
    # via jsii
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.101.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jsii
s3transfer==0.5.2
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   jsii
typing-extensions==4.12.2
    # via
    #   cattrs
    #   jsii
urllib3<2.0.0,>=1.26.0
    # via botocore
zipp==3.19.2
    # via importlib-resources
